requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
beautifulsoup4==4.9.3
boto3==1.17.34
botocore==1.20.34; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6; python_version >= '3.6'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-database-url==0.5.0
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-embed-video==1.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.11.1
django-taggit==1.3.0
django-templated-mail==1.1.1
django==3.1.7
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.2
djoser==2.1.0
gunicorn==20.0.4
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
itypes==1.2.0
jinja2==2.11.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pillow==8.1.2
psycopg2-binary==2.8.6
psycopg2==2.8.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==2.0.1; python_version >= '3.6'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
s3transfer==0.3.6
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==4.0.0
social-auth-core==4.1.0; python_version >= '3.6'
soupsieve==2.2.1; python_version >= '3.0'
sqlparse==0.4.1; python_version >= '3.5'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_full_version < '4.0.0'
whitenoise==5.2.0
